requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
pydantic>=2.0.0
pydantic[email]>=2.0.0
minio>=7.2.7
Pillow>=10.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0